+++ terraform/ecs-managed.tf
@@ -4,11 +4,75 @@ data "aws_vpc" "public" {
   }
 }
 
-data "aws_subnets" "public" {
-  // Get a list of subnets in the public vpc
+data "aws_subnet" "support" {
+  vpc_id = data.aws_vpc.public.id
   filter {
     name   = "vpc-id"
     values = [data.aws_vpc.public.id]
+  }
+  filter {
+    name   = "tag:Name"
+    values = ["Support Workstations"]
+  }
+}
+
+data "aws_subnet" "researcher" {
+  vpc_id = data.aws_vpc.public.id
+  filter {
+    name   = "vpc-id"
+    values = [data.aws_vpc.public.id]
+  }
+  filter {
+    name   = "tag:Name"
+    values = ["Researcher Workstations"]
+  }
+}
+
+data "aws_subnet" "three" {
+  vpc_id = data.aws_vpc.public.id
+  filter {
+    name   = "vpc-id"
+    values = [data.aws_vpc.public.id]
+  }
+  filter {
+    name   = "tag:Name"
+    values = ["Subnet 3"]
+  }
+}
+
+data "aws_subnet" "four" {
+  vpc_id = data.aws_vpc.public.id
+  filter {
+    name   = "vpc-id"
+    values = [data.aws_vpc.public.id]
+  }
+  filter {
+    name   = "tag:Name"
+    values = ["Subnet 4"]
+  }
+}
+
+data "aws_subnet" "five" {
+  vpc_id = data.aws_vpc.public.id
+  filter {
+    name   = "vpc-id"
+    values = [data.aws_vpc.public.id]
+  }
+  filter {
+    name   = "tag:Name"
+    values = ["Subnet 5"]
+  }
+}
+
+data "aws_subnet" "six" {
+  vpc_id = data.aws_vpc.public.id
+  filter {
+    name   = "vpc-id"
+    values = [data.aws_vpc.public.id]
+  }
+  filter {
+    name   = "tag:Name"
+    values = ["Subnet 6"]
   }
 }